// migrations/snapshots/516b08bb31dccca9578140e046205fa85f3e8827d6b44ef86c7f439f536e6949/contract.d.ts
export type CodecTypes = PgTypes;
export type LaneCodecTypes = CodecTypes;
export type QueryOperationTypes = PgTargetQueryOps<CodecTypes>;
// ...
export type FieldOutputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly diaVencimento: CodecTypes['pg/int4@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'] | null;
      readonly empresa: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly telefone: CodecTypes['pg/text@1']['output'] | null;
      readonly valorMensalidade: CodecTypes['pg/numeric@1']['output'] | null;
      readonly valorProduto: CodecTypes['pg/numeric@1']['output'] | null;
    };
    readonly Pagamento: {
      readonly clienteId: CodecTypes['pg/int4@1']['output'];
      readonly dataVencimento: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly diaVencimento: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly valor: CodecTypes['pg/numeric@1']['output'];
    };
    readonly Sessao: {
      readonly expiraEm: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly usuarioId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Tarefa: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly descricao: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly titulo: CodecTypes['pg/text@1']['output'];
    };
    readonly Usuario: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly perfil: 'ADMINISTRADOR' | 'FUNCIONARIO';
      readonly senha: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type FieldInputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly diaVencimento: CodecTypes['pg/int4@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'] | null;
      readonly empresa: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly telefone: CodecTypes['pg/text@1']['input'] | null;
      readonly valorMensalidade: CodecTypes['pg/numeric@1']['input'] | null;
      readonly valorProduto: CodecTypes['pg/numeric@1']['input'] | null;
    };
    readonly Pagamento: {
      readonly clienteId: CodecTypes['pg/int4@1']['input'];
      readonly dataVencimento: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly diaVencimento: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly valor: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Sessao: {
      readonly expiraEm: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly usuarioId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Tarefa: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly descricao: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly titulo: CodecTypes['pg/text@1']['input'];
    };
    readonly Usuario: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly perfil: 'ADMINISTRADOR' | 'FUNCIONARIO';
      readonly senha: CodecTypes['pg/text@1']['input'];
    };
  };
};
export type StorageColumnTypes = {
  readonly public: {
    readonly Cliente: {
      readonly diaVencimento: CodecTypes['pg/int4@1']['output'] | null;
      readonly email: CodecTypes['pg/text@1']['output'] | null;
      readonly empresa: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly telefone: CodecTypes['pg/text@1']['output'] | null;
      readonly valorMensalidade: CodecTypes['pg/numeric@1']['output'] | null;
      readonly valorProduto: CodecTypes['pg/numeric@1']['output'] | null;
    };
    readonly Pagamento: {
      readonly clienteId: CodecTypes['pg/int4@1']['output'];
      readonly dataVencimento: CodecTypes['pg/timestamptz-temporal@1']['output'] | null;
      readonly diaVencimento: CodecTypes['pg/int4@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly valor: CodecTypes['pg/numeric@1']['output'];
    };
    readonly Sessao: {
      readonly expiraEm: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly token: CodecTypes['pg/text@1']['output'];
      readonly usuarioId: CodecTypes['pg/int4@1']['output'];
    };
    readonly Tarefa: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly descricao: CodecTypes['pg/text@1']['output'] | null;
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly status: CodecTypes['pg/text@1']['output'];
      readonly titulo: CodecTypes['pg/text@1']['output'];
    };
    readonly Usuario: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['output'];
      readonly email: CodecTypes['pg/text@1']['output'];
      readonly id: CodecTypes['pg/int4@1']['output'];
      readonly nome: CodecTypes['pg/text@1']['output'];
      readonly perfil: 'ADMINISTRADOR' | 'FUNCIONARIO';
      readonly senha: CodecTypes['pg/text@1']['output'];
    };
  };
};
export type StorageColumnInputTypes = {
  readonly public: {
    readonly Cliente: {
      readonly diaVencimento: CodecTypes['pg/int4@1']['input'] | null;
      readonly email: CodecTypes['pg/text@1']['input'] | null;
      readonly empresa: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly telefone: CodecTypes['pg/text@1']['input'] | null;
      readonly valorMensalidade: CodecTypes['pg/numeric@1']['input'] | null;
      readonly valorProduto: CodecTypes['pg/numeric@1']['input'] | null;
    };
    readonly Pagamento: {
      readonly clienteId: CodecTypes['pg/int4@1']['input'];
      readonly dataVencimento: CodecTypes['pg/timestamptz-temporal@1']['input'] | null;
      readonly diaVencimento: CodecTypes['pg/int4@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly valor: CodecTypes['pg/numeric@1']['input'];
    };
    readonly Sessao: {
      readonly expiraEm: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly token: CodecTypes['pg/text@1']['input'];
      readonly usuarioId: CodecTypes['pg/int4@1']['input'];
    };
    readonly Tarefa: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly descricao: CodecTypes['pg/text@1']['input'] | null;
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly status: CodecTypes['pg/text@1']['input'];
      readonly titulo: CodecTypes['pg/text@1']['input'];
    };
    readonly Usuario: {
      readonly criadoEm: CodecTypes['pg/timestamptz-temporal@1']['input'];
      readonly email: CodecTypes['pg/text@1']['input'];
      readonly id: CodecTypes['pg/int4@1']['input'];
      readonly nome: CodecTypes['pg/text@1']['input'];
      readonly perfil: 'ADMINISTRADOR' | 'FUNCIONARIO';
      readonly senha: CodecTypes['pg/text@1']['input'];
    };
  };
};
// ...
export type TypeMaps = TypeMapsType<
  CodecTypes,
  QueryOperationTypes,
  FieldOutputTypes,
  FieldInputTypes,
  StorageColumnTypes,
  StorageColumnInputTypes,
  AggregateTypes
>;
// ...
export type Contract = ContractWithTypeMaps<ContractBase, TypeMaps>;